rename JdbcUrlParser to SimpleJdbcUrlParser
