+++ server/plugins/connect-plugin-api/src/main/java/com/oceanbase/odc/plugin/connect/api/JdbcUrlParser.java
@@ -13,10 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.core.shared.jdbc;
+package com.oceanbase.odc.plugin.connect.api;
 
 import java.util.List;
 import java.util.Map;
+
+import com.oceanbase.odc.core.shared.jdbc.HostAddress;
 
 /**
  * {@link JdbcUrlParser}
