+++ server/plugins/connect-plugin-ob-mysql/src/test/java/com/oceanbase/odc/plugin/connect/obmysql/OceanBaseJdbcUrlParserTest.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import com.oceanbase.odc.core.shared.jdbc.HostAddress;
-import com.oceanbase.odc.core.shared.jdbc.JdbcUrlParser;
+import com.oceanbase.odc.plugin.connect.api.JdbcUrlParser;
 
 /**
  * {@link OceanBaseJdbcUrlParserTest}
