+++ server/plugins/connect-plugin-oracle/src/test/java/com/oceanbase/odc/plugin/connect/oracle/OracleJdbcUrlParserTest.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import com.oceanbase.odc.core.shared.jdbc.HostAddress;
-import com.oceanbase.odc.core.shared.jdbc.JdbcUrlParser;
+import com.oceanbase.odc.plugin.connect.api.JdbcUrlParser;
 
 /**
  * @author jingtian
